requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.1.3
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==42.0.7
    # via pyjwt
dnspython==2.6.1
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-cli
fastapi-cli==0.0.2
    # via fastapi
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.3
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.3
    # via fastapi
passlib==1.7.4
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
pygments==2.17.2
    # via rich
pyjwt[crypto]==2.8.0
    # via -r requirements.in
pyodbc==5.1.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
ujson==5.9.0
    # via fastapi
uvicorn[standard]==0.29.0
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
